+++ src/task-edit.js
@@ -1,8 +1,9 @@
-import {createElement} from './utils/index.js';
+import {Component} from './component.js';
 import {getRandomBoolean, COLOR_TASKS, REPEATING_DAYS} from './utils/index.js';
 
-export class TaskEdit {
+class TaskEdit extends Component {
   constructor(data) {
+    super();
     this._title = data.title;
     this._dueDate = data.dueDate;
     this._picture = data.picture;
@@ -12,7 +13,6 @@ export class TaskEdit {
     this._isFavorite = data.isFavorite;
     this._isDone = data.isDone;
 
-    this._element = null;
     this._onSubmit = null;
     this._onDelete = null;
     this._onKeyEsc = null;
@@ -20,10 +20,6 @@ export class TaskEdit {
     this._onSubmitButtonClick = this._onSubmitButtonClick.bind(this);
     this._onDeleteButtonClick = this._onDeleteButtonClick.bind(this);
     this._onKeydownEsc = this._onKeydownEsc.bind(this);
-  }
-
-  get element() {
-    return this._element;
   }
 
   _onSubmitButtonClick(evt) {
@@ -92,14 +88,6 @@ export class TaskEdit {
       }
     }
     return htmlRepeat;
-  }
-
-  _getDate(time) {
-    return new Date(time).toLocaleString(`en-US`, {year: `numeric`, month: `long`, day: `numeric`});
-  }
-
-  _getTime(time) {
-    return new Date(time).toLocaleString(`en-US`, {hour: `numeric`, minute: `numeric`});
   }
 
   get template() {
@@ -187,16 +175,6 @@ export class TaskEdit {
       .removeEventListener(`click`, this._onDeleteButtonClick);
     document.removeEventListener(`keydown`, this._onKeydownEsc);
   }
-
-  render() {
-    this._element = createElement(this.template);
-    this.bind();
-
-    return this._element;
-  }
-
-  unrender() {
-    this.unbind();
-    this._element = null;
-  }
 }
+
+export {TaskEdit};
